requirements.txt
# Core deep learning
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# Data handling / utils
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0
pyyaml==6.0.2

# Progress bar / logging
tqdm==4.66.4
